Proyecto

+++ src/main/kotlin/com/example/Restaurantes/Model/Restaurantes.kt
@@ -14,5 +14,5 @@ class Restaurantes {
     var nombre: String? = null
     var calificacion: Int? = null
     var direccion: String? = null
-    var duenos_iddueno: Int? = null
+    var duenos_iddueno: Long? = null
 }
